Get rid of spinner
