+++ package.json
@@ -20,11 +20,10 @@
     "react-moment": "^1.1.2",
     "react-router-dom": "^6.4.2",
     "react-scripts": "5.0.1",
-    "react-spinners-css": "^2.0.1",
     "web-vitals": "^2.1.0"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "GENERATE_SOURCEMAP=false && react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
